chore: move lifecycle handler in separate file
