+++ mobile/lib/pages/common/native_video_viewer.page.dart
@@ -3,7 +3,6 @@ import 'dart:io';
 
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter_hooks/flutter_hooks.dart' hide Store;
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/domain/models/store.model.dart';
@@ -19,6 +18,7 @@ import 'package:immich_mobile/services/app_settings.service.dart';
 import 'package:immich_mobile/services/asset.service.dart';
 import 'package:immich_mobile/utils/debounce.dart';
 import 'package:immich_mobile/utils/hooks/interval_hook.dart';
+import 'package:immich_mobile/utils/lifecycle_handler.dart';
 import 'package:immich_mobile/widgets/asset_viewer/custom_video_player_controls.dart';
 import 'package:logging/logging.dart';
 import 'package:native_video_player/native_video_player.dart';
@@ -423,30 +423,5 @@ class NativeVideoViewerPage extends HookConsumerWidget {
         if (showControls) const Center(child: CustomVideoPlayerControls()),
       ],
     );
-  }
-}
-
-class LifecycleEventHandler extends WidgetsBindingObserver {
-  final AsyncCallback? onResume;
-  final AsyncCallback? onPause;
-
-  LifecycleEventHandler({this.onResume, this.onPause});
-
-  @override
-  void didChangeAppLifecycleState(AppLifecycleState state) {
-    switch (state) {
-      case AppLifecycleState.resumed:
-        if (onResume != null) {
-          onResume!();
-        }
-        break;
-      case AppLifecycleState.paused:
-        if (onPause != null) {
-          onPause!();
-        }
-        break;
-      default:
-        break;
-    }
   }
 }
